split up ShiftHighAndLowCommandTests into ShiftHighCommandTest and ShiftLowCommandTest
