+++ Competition2018/src/test/java/competition/subsystems/shift/ShiftLowCommandTest.java
@@ -1,18 +1,15 @@
 package competition.subsystems.shift;
 
 import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
 
 import competition.BaseCompetitionTest;
 import competition.subsystems.shift.ShiftSubsystem.Gear;
-import competition.subsystems.shift.commands.ShiftHighCommand;
 import competition.subsystems.shift.commands.ShiftLowCommand;
 import edu.wpi.first.wpilibj.MockSolenoid;
 
-public class ShiftHighAndLowCommandsTest extends BaseCompetitionTest {
-    ShiftHighCommand shiftHighCommand;
+public class ShiftLowCommandTest extends BaseCompetitionTest {
     ShiftLowCommand shiftLowCommand;
     ShiftSubsystem subsystem;
     MockSolenoid solenoid;
@@ -20,26 +17,11 @@ public class ShiftHighAndLowCommandsTest extends BaseCompetitionTest {
     @Override
     public void setUp() {
         super.setUp();
-
-        shiftHighCommand = injector.getInstance(ShiftHighCommand.class);
         shiftLowCommand = injector.getInstance(ShiftLowCommand.class);
         subsystem = injector.getInstance(ShiftSubsystem.class);
-
         solenoid = ((MockSolenoid) subsystem.solenoid);
     }
-
-    @Test
-    public void testShiftHighCommand() {
-        // shift from low to high
-        subsystem.setGear(Gear.LOW_GEAR);
-        shiftHighCommand.initialize();
-        assertTrue(solenoid.get());
-
-        // shift from high to high
-        shiftHighCommand.initialize();
-        assertTrue(solenoid.get());
-    }
-
+    
     @Test
     public void testShiftLowCommand() {
         // shift from high to low
@@ -51,5 +33,4 @@ public class ShiftHighAndLowCommandsTest extends BaseCompetitionTest {
         shiftLowCommand.initialize();
         assertFalse(solenoid.get());
     }
-
 }
